+++ client/src/router/main.tsx
@@ -2,20 +2,22 @@ import React from "react";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
 import Register from "../pages/Reg";
+import Settings from "../pages/Settings";
 
 import { Navigate } from "react-router-dom";
 
 export const privateRoutes: RouteT[] = [
-  // { path: "/posts", component: <Posts /> },
+  { path: "/settings/general", component: <Settings /> },
   // { path: "/posts/:id", component: <PostId /> },
-  // { path: "*", component: <Navigate to="/posts" /> },
+  { path: "/settings", component: <Navigate to="/settings/general" /> },
+  // { path: "*", component: <Navigate to="/" /> },
 ];
 
 export const publicRoutes: RouteT[] = [
   { path: "/", component: <Home /> },
   { path: "/login", component: <Login /> },
   { path: "/reg", component: <Register /> },
-  // { path: "*", component: <Navigate to="/login" /> },
+  { path: "*", component: <Navigate to="/login" /> },
 ];
 
 type RouteT = { path: string; component: React.ReactNode };