+++ server/src/db/models.ts
@@ -1,3 +1,7 @@
 import User from "./models/user.model";
+import Transfer from "./models/transfer.model";
 
-export const models = { User };
+export const models = {
+  User,
+  Transfer,
+};